+++ CHANGES.md
@@ -21,6 +21,7 @@ Changes in 10.0
 - Add User photo tooltip in Grades/TeacherCompletion.php
 - SQL order by Period title in TeacherCompletion.php
 - Use Period's Short Name when > 10 columns in the list in TeacherCompletion.php
+- Fix SQL to select Periods where exists CP in TeacherCompletion.php & Administration.php
 
 Changes in 9.3
 --------------